# [Enhancement](function) optimize for decimal arithmetic calculation